🚚build: 파일 그룹화 및 storyboard 삭제
