+++ Diary/App/SceneDelegate.swift
@@ -1,0 +1,38 @@
+//
+//  Diary - SceneDelegate.swift
+//  Created by Dasan, kyungmin on 2023/08/28.
+//  Copyright © yagom. All rights reserved.
+// 
+
+import UIKit
+
+class SceneDelegate: UIResponder, UIWindowSceneDelegate {
+
+    var window: UIWindow?
+
+
+    func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
+        guard let windowSecene = (scene as? UIWindowScene) else {
+            return
+        }
+        
+        window = UIWindow(windowScene: windowSecene)
+        let viewController = ViewController()
+        
+        window?.rootViewController = UINavigationController(rootViewController: viewController)
+        window?.makeKeyAndVisible()
+    }
+
+    func sceneDidDisconnect(_ scene: UIScene) {}
+
+    func sceneDidBecomeActive(_ scene: UIScene) {}
+
+    func sceneWillResignActive(_ scene: UIScene) {}
+
+    func sceneWillEnterForeground(_ scene: UIScene) {}
+
+    func sceneDidEnterBackground(_ scene: UIScene) {
+        (UIApplication.shared.delegate as? AppDelegate)?.saveContext()
+    }
+}
+
